Move logAdminActivity enums to module
