+++ pages/api/id/[form]/owners.ts
@@ -5,10 +5,9 @@ import { Prisma } from "@prisma/client";
 import { isValidGovEmail } from "@lib/validation";
 import emailDomainList from "../../../../email.domains.json";
 import { Session } from "next-auth";
-import { logAdminActivity } from "@lib/adminLogs";
+import { logAdminActivity, AdminLogAction, AdminLogEvent } from "@lib/adminLogs";
 import { logMessage } from "@lib/logger";
 import { MiddlewareProps } from "@lib/types";
-import { AdminLogAction, AdminLogEvent } from "@lib/types/utility-types";
 
 const handler = async (
   req: NextApiRequest,
